+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Validate iterator chunks passed to `Utils::streamFor()`
 - Stop adding default `Content-Length` headers to `multipart/form-data` parts to comply with RFC 7578 section 4.8
 
+### Fixed
+
+- Do not treat objects with `__call()` as stringable in `Utils::streamFor()`
+
 ## 2.10.0 - 2026-05-19
 
 ### Changed